requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.4.1
autopep8==1.6.0
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.9; python_version >= '3'
cloudinary==1.28.0
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
django-bootstrap4==3.0.1
django-cors-headers==3.10.1
django-crispy-forms==1.13.0
django-heroku==0.3.1
django==3.2.8
djangorestframework==3.13.1
drf-flex-fields==0.9.6
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.3; python_version >= '3'
inflection==0.5.1; python_version >= '3.5'
itypes==1.2.0
jinja2==3.0.3; python_version >= '3.6'
markupsafe==2.0.1; python_version >= '3.6'
packaging==21.3; python_version >= '3.6'
pillow==8.3.2
psycopg2-binary==2.9.2
psycopg2==2.9.1
pycodestyle==2.8.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
pycryptodome==3.12.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
pyparsing==3.0.6; python_version >= '3.6'
python-decouple==3.5
python-dotenv==0.19.2
pytz==2021.3
rave-python==1.2.15
requests==2.26.0
ruamel.yaml.clib==0.2.6; python_version < '3.10' and platform_python_implementation == 'CPython'
ruamel.yaml==0.17.17; python_version >= '3'
setuptools==59.7.0; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
soupsieve==2.2.1
sqlparse==0.4.2
toml==0.10.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
uritemplate==4.1.1; python_version >= '3.6'
urllib3==1.26.7; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
whitenoise==5.3.0
